+++ web/src/components/organisms/index.ts
@@ -1,3 +1,6 @@
 export * from './SearchControlsPanel';
 export * from './RetrievalResultsPanel';
 export * from './ResultInspectorPanel';
+export * from './QueueHealthPanel';
+export * from './WorkflowListPanel';
+export * from './CorpusSourcePanel';